rename integer_utils
