+++ src/util/src/algebra_utils.rs
