+++ src/util/src/prime.rs
@@ -20,7 +20,7 @@ use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use crate::biguint_serde;
 use crate::{
     csprng::Csprng,
-    integer_util::{cnt_bits_repr_usize, largest_integer_a_such_that_2_to_a_divides_even_n},
+    algebra_utils::{cnt_bits_repr_usize, largest_integer_a_such_that_2_to_a_divides_even_n},
 };
 
 pub const PRIMES_TABLE_U8: [u8; 54] = [
